+++ athena/athena/models/__init__.py
@@ -5,14 +5,14 @@
 
 from .db_exercise import DBExercise
 from .db_submission import DBSubmission
-from athena.models.feedback.graded_feedback.db_graded_feedback import DBGradedFeedback
+from .db_graded_feedback import DBGradedFeedback
 from .db_programming_exercise import DBProgrammingExercise
 from .db_text_exercise import DBTextExercise
 from .db_modelling_exercise import DBModellingExercise
 from .db_programming_submission import DBProgrammingSubmission
 from .db_text_submission import DBTextSubmission
 from .db_modelling_submission import DBModellingSubmission
-from .feedback import DBGradedProgrammingFeedback
-from .feedback import DBNonGradedProgrammingFeedback
-from .feedback import DBGradedTextFeedback
-from .feedback import DBGradedModellingFeedback
+from .db_graded_programming_feedback import DBGradedProgrammingFeedback
+from .db_non_graded_programming_feedback import DBNonGradedProgrammingFeedback
+from .db_graded_text_feedback import DBGradedTextFeedback
+from .db_graded_modelling_feedback import DBGradedModellingFeedback
