+++ athena/athena/models/db_graded_programming_feedback.py
@@ -3,12 +3,12 @@ from sqlalchemy import Column, Integer, String, ForeignKey
 from sqlalchemy.orm import relationship
 
 from athena.database import Base
-from athena.models.feedback.non_graded_feedback.db_non_graded_feedback import DBNonGradedFeedback
-from athena.models.big_integer_with_autoincrement import BigIntegerWithAutoincrement
+from .db_graded_feedback import DBGradedFeedback
+from .big_integer_with_autoincrement import BigIntegerWithAutoincrement
 
 
-class DBNonGradedProgrammingFeedback(DBNonGradedFeedback, Base):
-    __tablename__ = "non_graded_programming_feedbacks"
+class DBGradedProgrammingFeedback(DBGradedFeedback, Base):
+    __tablename__ = "graded_programming_feedbacks"
 
     file_path: Optional[str] = Column(String)  # type: ignore
     line_start: Optional[int] = Column(Integer)  # type: ignore
@@ -17,5 +17,5 @@ class DBNonGradedProgrammingFeedback(DBNonGradedFeedback, Base):
     exercise_id = Column(BigIntegerWithAutoincrement, ForeignKey("programming_exercises.id", ondelete="CASCADE"), index=True)
     submission_id = Column(BigIntegerWithAutoincrement, ForeignKey("programming_submissions.id", ondelete="CASCADE"), index=True)
 
-    exercise = relationship("DBProgrammingExercise", back_populates="non_graded_feedbacks")
-    submission = relationship("DBProgrammingSubmission", back_populates="non_graded_feedbacks")
+    exercise = relationship("DBProgrammingExercise", back_populates="graded_feedbacks")
+    submission = relationship("DBProgrammingSubmission", back_populates="graded_feedbacks")
