+++ athena/athena/models/db_non_graded_feedback.py
@@ -1,0 +1,5 @@
+from .db_feedback import DBFeedback
+
+
+class DBNonGradedFeedback(DBFeedback):
+    pass
