--- athena/athena/models/feedback/__init__.py
@@ -1,7 +1,0 @@
-from .db_feedback import DBFeedback
-from .graded_feedback import DBGradedFeedback
-from .graded_feedback import DBGradedModellingFeedback
-from .graded_feedback import DBGradedProgrammingFeedback
-from .graded_feedback import DBGradedTextFeedback
-from .non_graded_feedback import DBNonGradedFeedback
-from .non_graded_feedback import DBNonGradedProgrammingFeedback
