--- athena/athena/models/feedback/graded_feedback/__init__.py
@@ -1,4 +1,0 @@
-from .db_graded_feedback import DBGradedFeedback
-from .db_graded_modelling_feedback import DBGradedModellingFeedback
-from .db_graded_programming_feedback import DBGradedProgrammingFeedback
-from .db_graded_text_feedback import DBGradedTextFeedback
