--- athena/athena/models/feedback/non_graded_feedback/__init__.py
@@ -1,2 +1,0 @@
-from .db_non_graded_feedback import DBNonGradedFeedback
-from .db_non_graded_programming_feedback import DBNonGradedProgrammingFeedback
