--- athena/athena/models/feedback/non_graded_feedback/db_non_graded_feedback.py
@@ -1,5 +1,0 @@
-from athena.models.feedback.db_feedback import DBFeedback
-
-
-class DBNonGradedFeedback(DBFeedback):
-    pass
